Rename TextInput to Input
